requirements.txt
python-docx>=1.2.0
rich>=13.0.0
google-genai>=0.2.0
